requirements.txt
PyJWT[crypto]>=2.1.0
cryptography>=3.3.2
requests>=2.25.1
python-dotenv>=0.15.0